+++ src/routes/settings/content/+page.svelte
@@ -87,36 +87,6 @@
         </div>
       {/snippet}
     </Field>
-    <strong class="text-lg">Privacy Settings</strong>
-    <FieldInline>
-      {#snippet label()}
-        <p>Report errors?</p>
-      {/snippet}
-      {#snippet input()}
-        <input
-          type="checkbox"
-          class="toggle toggle-primary"
-          bind:checked={settings.report_errors} />
-      {/snippet}
-      {#snippet info()}
-        <p>
-          Allow {PLATFORM_NAME} to send error reports to help improve the app.
-        </p>
-      {/snippet}
-    </FieldInline>
-    <FieldInline>
-      {#snippet label()}
-        <p>Report usage?</p>
-      {/snippet}
-      {#snippet input()}
-        <input type="checkbox" class="toggle toggle-primary" bind:checked={settings.report_usage} />
-      {/snippet}
-      {#snippet info()}
-        <p>
-          Allow {PLATFORM_NAME} to collect anonymous usage data.
-        </p>
-      {/snippet}
-    </FieldInline>
     <strong class="text-lg">Editor Settings</strong>
     <FieldInline>
       {#snippet label()}